Remember initial TP with session ticket

+++ picoquic/picoquic.h
@@ -268,12 +268,12 @@ typedef struct st_picoquic_tp_prefered_address_t {
 } picoquic_tp_prefered_address_t;
 
 typedef struct st_picoquic_tp_t {
-    uint32_t initial_max_stream_data_bidi_local;
-    uint32_t initial_max_stream_data_bidi_remote;
-    uint32_t initial_max_stream_data_uni;
-    uint32_t initial_max_data;
-    uint32_t initial_max_stream_id_bidir;
-    uint32_t initial_max_stream_id_unidir;
+    uint64_t initial_max_stream_data_bidi_local;
+    uint64_t initial_max_stream_data_bidi_remote;
+    uint64_t initial_max_stream_data_uni;
+    uint64_t initial_max_data;
+    uint64_t initial_max_stream_id_bidir;
+    uint64_t initial_max_stream_id_unidir;
     uint32_t idle_timeout;
     uint32_t max_packet_size;
     uint32_t max_ack_delay; /* stored in in microseconds for convenience */
